requirements.txt
openai
requests
python-dotenv
matplotlib
textblob
